Clean up evaluate script a bit
