+++ elk/evaluate.py
@@ -1,7 +1,7 @@
-from .training.ccs import CCS
-from .eval.utils_evaluation import load_hidden_states
-from .eval.parser import get_args
 from .files import elk_cache_dir
+from .training.ccs import CCS
+from .training.preprocessing import load_hidden_states
+from argparse import ArgumentParser
 from sklearn.model_selection import train_test_split
 import pickle
 import torch
@@ -25,13 +25,26 @@ def evaluate(args, hiddens, labels, lr_model, ccs_model: CCS):
     print(f"accuracy_lr {acc_lr}")
 
 
-if __name__ == "__main__":
-    args = get_args()
-
-    hiddens, labels = load_hidden_states(
-        path=elk_cache_dir() / args.name / "hiddens.pt",
-        reduce=args.mode,
+def main():
+    parser = ArgumentParser()
+    parser.add_argument("name", type=str, help="Name of the experiment")
+    parser.add_argument(
+        "--transfer-to",
+        type=str,
+        help="Name of experiment whose hidden states to evaluate on.",
     )
+    parser.add_argument("--device", type=str, help="PyTorch device to use.")
+    parser.add_argument("--layer", type=int, default=-1)
+    parser.add_argument("--seed", type=int, default=0)
+    args = parser.parse_args()
+
+    # Default to CUDA iff available
+    if args.device is None:
+        args.device = "cuda" if torch.cuda.is_available() else "cpu"
+
+    hidden_path = elk_cache_dir() / (args.transfer_to or args.name)
+    print(f"Loading hidden states from \033[1m{hidden_path}\033[0m")  # bold
+    hiddens, labels = load_hidden_states(path=hidden_path / "hiddens.pt")
     _, test_hiddens, __, test_labels = train_test_split(
         hiddens, labels, random_state=args.seed, stratify=labels
     )
@@ -43,3 +56,7 @@ if __name__ == "__main__":
     ccs_models = torch.load(path / "ccs_models.pt")
     for h, lr_model, ccs_model in zip(test_hiddens.unbind(1), lr_models, ccs_models):
         evaluate(args, h, test_labels, lr_model, ccs_model)
+
+
+if __name__ == "__main__":
+    main()
